backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
twilio==8.2.2
earthengine-api==0.1.374
tensorflow==2.14.0
scikit-learn==1.3.2
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
joblib==1.3.2
openai-whisper==20231117
gtts==2.5.0
speechrecognition==3.10.0
pydub==0.25.1
folium==0.15.1
geopy==2.4.1
requests==2.31.0